Hide and show the LutrisWindow via the tray icon
Clicking it will toggle the visibility of the LutrisWindow, This is a bit trickier than it seems though; we need tor restore its position as that gets reset, and therefore we must save it when hiding.

Further, hiding a window that has a transient dialog breaks GTK
messily, so I don't allow that.

Resolves #4352