Try harder to find the game_path for Linux games
This code can find the game in '/usr/games' or the like. This allows
'browse files' to work for those games, but more importantly this makes
the 'beat' code that watches for the game to exit more accurate.

Without this, it considers every new process to be part of the game, so
it likely never sees the game as having exited.

With this, it's not completely precise but it's more accurate at least.