Sort the versions in the WINE Manage Versions dialog
The code looks for a version number it can parse
from the end of the version string.

It parses the version into a list of ints, but does
treat - and . and equivalent, which might be a problem in future.

When no version can be extracted,
we'll sort by the whole version string.